dom/workers/test/test_url.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<head>
  <title>Test for URL object in workers</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<pre id="test"></pre>
<script class="testbody" type="text/javascript">

  var worker = new Worker("url_worker.js");

  worker.onmessage = function(event) {
    is(event.target, worker);

    if (event.data.type == 'finish') {
      runTest();
    } else if (event.data.type == 'status') {
      ok(event.data.status, event.data.msg);
    } else if (event.data.type == 'url') {
      var xhr = new XMLHttpRequest();
      xhr.open('GET', event.data.url, false);
      xhr.onreadystatechange = function() {
        if (xhr.readyState == 4) {
          ok(true, "Blob readable!");
        }
      }
      xhr.send();
    }
  };

  worker.onerror = function(event) {
    is(event.target, worker);
    ok(false, "Worker had an error: " + event.data);
    SimpleTest.finish();
  };

  var tests = [
    function() { worker.postMessage(0); },
    function() { worker.postMessage(1); }
  ];

  function runTest() {
    if (!tests.length) {
      SimpleTest.finish();
      return;
    }

    var test = tests.shift();
    test();
  }

  runTest();
  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
